--- .qualified-attach.json
@@ -1,1 +1,0 @@
-{"token":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ0b2tlbl90eXBlIjoiYXR0YWNoIiwiY2FuZGlkYXRlX2lkIjoiNjUwZGRlYTNiZThlNjUwMDEzZGM2M2MxIiwiYXNzZXNzbWVudF9yZXN1bHRfaWQiOiI2NThkY2E3ZGQ0NWUwMjM1MDFkN2VkMTkiLCJjaGFsbGVuZ2VfaWQiOiI1ZjljNDVjNTcwZTA1MTAwMGEzYzAwYWIifQ.eOIhFJxwnlf9-q7hOQ-cP1EbbRtiAchktxZgAVoJyhc"}
